Module tabin_plugins::procfs

Structs and impls for the various files from the /proc filesystem

Each file gets a struct to represent its data, with an associated load function.

Modules

pid
Data structures related to the /proc/<pid>/* files

Structs

Calculations
The number of calculations that have occured on this computer in a time period

InvalidWorkSource

LoadAvg
The load average of the system

LoadProcsError

MemInfo
A struct that represents overall memory usage on the system.

Mount

ParseStatError

ParseStateError

ProcessCpuUsages
A collection of ProcessCpuUsages

RunningProcs
All the processes that are running

Enums

ProcField

ProcFsError
ProcFs errors

WorkSource
A kind of CPU usage

Type Definitions

ProcMap

Result
All the results are results with ProcFsErrors