Enum tabin_plugins::procfs::ProcFsError

pub enum ProcFsError {
    Io(Error),
    InsufficientData(String),
    InvalidFloat(ParseFloatError),
    InvalidInt(ParseIntError),
    LoadProcsError(LoadProcsError),
    ParseStatError(ParseStatError),
    ParseStateError(ParseStateError),
}

ProcFs errors

Every error from in this module can be converted into a ProcFsError

Variants
Io(Error)

Errors originating in IO

InsufficientData(String)

Error pulling all required data out of procfs

InvalidFloat(ParseFloatError)

Happens when we try to parse a float from something in procfs

InvalidInt(ParseIntError)

Happens when we try to parse an int from something in procfs

LoadProcsError(LoadProcsError)

When we receive an error loading an individual proc, we get this

ParseStatError(ParseStatError)

Happens when we try to parse a line from /proc//stat and got an error

ParseStateError(ParseStateError)

Happens when we get an invalid process state