Struct tabin_plugins::procfs::pid::Process

pub struct Process {
    pub stat: Stat,
    pub cmdline: CmdLine,
}

Information about a running process

Fields
stat: Stat

The stat info for a process

cmdline: CmdLine

The command line, as revealed by the /proc fs

Methods

impl Process

pub fn from_pid<P: Display + Copy>(p: P) -> Result<Process>

pub fn useful_cmdline(&self) -> String

pub fn percent_ram(&self, of_bytes: usize) -> f64

What percent this process is using

First argument should be in bytes.

pub fn cpu_utilization_since<'a>(
&'a self,
start_process: &'a Process,
total_cpu: Jiffies
) -> ProcessCpuUsage<'a>

Compare this processes cpu utilization since the process representing th start time

The passed-in start_process is the time that we are comparing against: self should be newer.

The total_cpu is how many Jiffies have passed on the cpu over the same time period.

Panics

If the total_cpu on start_process is higher than the total_cpu on current.