Struct tabin_plugins::procfs::Calculations

pub struct Calculations {
    pub user: Jiffies,
    pub nice: Jiffies,
    pub system: Jiffies,
    pub idle: Jiffies,
    pub iowait: Jiffies,
    pub irq: Jiffies,
    pub softirq: Jiffies,
    pub steal: Jiffies,
    pub guest: Jiffies,
    pub guest_nice: Option<Jiffies>,
}

The number of calculations that have occured on this computer in a time period

Fields

user

Time spent in user mode.

nice

Time spent in user mode with low priority (nice).

system

Time spent in system mode.

idle

Time spent not doing anything in particular

iowait

Time forced to be idle waiting for I/O to complete.

irq

Time spent servicing hardware interrupts

softirq

Time spent servicing software interrupts

steal

Stolen time, which is the time spent in other operating systems when running in a virtualized environment

guest

Time spent running a virtual CPU for a guest operating system. AKA time we lent out to virtual machines.

guest_nice

Time spent running a niced guest

Methods

impl Calculations

fn load() -> Result<Calculations>

Build a new Calculations for total CPU jiffies from the /proc/stat pseudofile. See load_per_cpu for per-cpu metrics.

fn load_per_cpu() -> Result<Vec<Calculations>>

Build a list of per-cpu metrics from the /proc/stat file

This does not include the total line, use load() for that.

fn active(&self) -> Jiffies

Jiffies spent non-idle

This includes all processes in user space, kernel space, and time stolen by other VMs.

fn idle(&self) -> Jiffies

Jiffies spent with nothing to do

fn virt(&self) -> Jiffies

Jiffies spent running child VMs

This is included in active() by the nature of the way that /proc/stats reports things, so don't add this to that when totalling.

fn total(&self) -> Jiffies

All jiffies since the kernel started tracking

fn percent_util_since(&self, kind: &WorkSource, start: &Calculations) -> f64

Return how much cpu time the specific worksource took

The number returned is between 0 and 100