Struct tabin_plugins::procfs::MemInfo

pub struct MemInfo {
    pub total: Option<usize>,
    pub available: Option<usize>,
    pub free: Option<usize>,
    pub cached: Option<usize>,
}

A struct that represents overall memory usage on the system.

All values are in KB.

Fields
total: Option<usize> available: Option<usize> free: Option<usize> cached: Option<usize>
Methods

impl MemInfo

pub fn load() -> MemInfo

Read the data from /proc/meminfo into a MemInfo

pub fn percent_free(&self) -> Result<f64>

Try to figure out how much memory is being used

Since linux kernel 3.16 this just performs

    available / total * 100

Before that we approximate available as Free + Cached, even though that is almost certain to be incorrect.

pub fn percent_used(&self) -> Result<f64>

The inverse of MemInfo::percent_free