Struct tabin_plugins::procfs::RunningProcs

pub struct RunningProcs(pub ProcMap);

All the processes that are running

Methods

impl RunningProcs

pub fn currently_running() -> Result<RunningProcs>

Load the currently running processes from /proc/[pid]/*

pub fn percent_cpu_util_since<'a>(
&self,
start: &'a RunningProcs,
total_cpu: Jiffies
) -> ProcUsages<'a>

Collect CPU usage for each proc as compared to the total active CPU time for the system.

The value for total_cpu should probably be the result of subtracting two Calculations::total()s from each other.