Struct tabin_plugins::sys::fs::cgroup::memory::Stat

pub struct Stat {
    pub cache: usize,
    pub rss: usize,
    pub rss_huge: usize,
    pub swap: usize,
}

Some fields from the memory.stat file

All values are in bytes, contrary to the same file from procfs which reports everything in pages

Fields
cache: usize

Memory used, including the filesystem page cache. This number will never decrease unless memory presure gets applied from outside the cgroup.

rss: usize

Actual memory being used by the cgroup

rss_huge: usize

Only hugetables. This is included in rss

swap: usize

Number of bytes that have been swapped out to disk

Methods

impl Stat

pub fn load() -> Result<Stat, Error>

Read information from the filesystem and create a new Stat