Crate tabin_plugins

Utitilty for writing nagios-style check scripts/plugins

There are three things:

The Status enum for representing health status

The procfs module, which contains rusty representations of some files from /proc

A few scripts in the bin directory, which contain actual nagios-compatible scripts

TODOs include

Nice logging, including some standard formats -- json and human would be nice

Some way of easily standardizing command-line args

Much of the code is hideous, and should not be

Modules

linux
Standard APIs in Linux

procfs
Structs and impls for the various files from the /proc filesystem

scripts
Documentation about the various scripts contained herein

sys
Interact with the /sys pseudo-file system

Enums

Status
Represent the nagios-ish error status of a script.

TurbineError
All errors are TurbineErrors

Type Definitions

TurbineResult
All results are TurbineResults