Module tabin_plugins::procfs

Structs and impls for the various files from the /proc filesystem

Each file gets a struct to represent its data, with an associated load function.

Modules

pid
Data structures related to the /proc//* files

Structs

Calculations
The number of calculations that have occured on this computer in a time period

LoadAvg
The load average of the system

MemInfo
A struct that represents overall memory usage on the system.

Mount

ProcUsage
Represent the percent CPU utilization of a specific process over a specific time period

ProcUsages

RunningProcs
All the processes that are running

Enums

ProcFsError
ProcFs errors

WorkSource
A kind of CPU usage

Type Definitions

ProcMap

Result
All the results are results with ProcFsErrors

Usages